move functions into helpers
