+++ support-frontend/assets/helpers/tracking/quantumMetricHelpers.ts
@@ -1,11 +1,13 @@
+import { onConsentChange } from '@guardian/consent-management-platform';
+import { isSwitchOn } from 'helpers/globalsAndSwitches/globals';
 import type { BillingPeriod } from 'helpers/productPrice/billingPeriods';
 import type { ProductPrice } from 'helpers/productPrice/productPrices';
 import { getAppliedPromo } from 'helpers/productPrice/promotions';
 
-export const getSubscriptionAnnualValue = (
+export function getSubscriptionAnnualValue(
 	productPrice: ProductPrice,
 	billingPeriod: BillingPeriod,
-): number | undefined => {
+): number | undefined {
 	const fullPrice = productPrice.price;
 	const promotion = getAppliedPromo(productPrice.promotions);
 
@@ -44,4 +46,42 @@ export const getSubscriptionAnnualValue = (
 		discountInPenceCents * promotion.numberOfDiscountedPeriods;
 
 	return discountedAnnualPrice;
-};
+}
+
+export function waitForQuantumMetricAPi(onReady: () => void): void {
+	let pollCount = 0;
+	const checkForQuantumMetricAPi = setInterval(() => {
+		pollCount = pollCount + 1;
+		if (window.QuantumMetricAPI?.isOn()) {
+			onReady();
+			clearInterval(checkForQuantumMetricAPi);
+		} else if (pollCount === 10) {
+			// give up waiting if QuantumMetricAPI is not ready after 10 attempts
+			clearInterval(checkForQuantumMetricAPi);
+		}
+	}, 500);
+}
+
+export function canRunQuantumMetric(): Promise<boolean> {
+	// resolve immediately with false if the feature switch is OFF
+	if (!isSwitchOn('featureSwitches.enableQuantumMetric')) {
+		return Promise.resolve(false);
+	}
+	// checks users consent status
+	return new Promise((resolve) => {
+		onConsentChange((state) => {
+			if (
+				state.ccpa?.doNotSell === false || // check whether US users have NOT withdrawn consent
+				state.aus?.personalisedAdvertising || // check whether AUS users have consented to personalisedAdvertising
+				(state.tcfv2?.consents && // check TCFv2 purposes for non-US/AUS users
+					state.tcfv2.consents['1'] && // Store and/or access information on a device
+					state.tcfv2.consents['8'] && // Measure content performance
+					state.tcfv2.consents['10']) // Develop and improve products
+			) {
+				resolve(true);
+			} else {
+				resolve(false);
+			}
+		});
+	});
+}
